pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "offshore"
version = "0.1.0"
description = "Single-stock trend classification using state-of-the-art time-series deep learning"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.11"
authors = [{ name = "Offshore Team", email = "offshore@example.com" }]
keywords = [
    "stock prediction",
    "time series",
    "deep learning",
    "transformer",
    "lstm",
    "trend classification",
    "pytorch",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Office/Business :: Financial :: Investment",
]

dependencies = [
    "torch>=2.0.0",
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "scikit-learn>=1.3.0",
    "yfinance>=0.2.28",
    "pyyaml>=6.0",
    "typer>=0.9.0",
    "rich>=13.0.0",
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
    "tqdm>=4.65.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "mypy>=1.5.0",
    "black>=23.7.0",
    "isort>=5.12.0",
    "flake8>=6.1.0",
    "pre-commit>=3.3.0",
]
notebooks = ["jupyter>=1.0.0", "jupyterlab>=4.0.0", "ipywidgets>=8.1.0"]

[project.scripts]
offshore = "src.cli.main:app"

[project.urls]
Homepage = "https://github.com/yourusername/offshore"
Documentation = "https://github.com/yourusername/offshore#readme"
Repository = "https://github.com/yourusername/offshore"
Issues = "https://github.com/yourusername/offshore/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.setuptools.package-data]
"*" = ["*.yaml", "*.yml"]

[tool.black]
line-length = 100
target-version = ['py311']
include = '\.pyi?$'
exclude = '''
/(
    \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
  | data
)/
'''

[tool.isort]
profile = "black"
line_length = 100
skip = [".gitignore", ".dockerignore"]
skip_glob = ["data/*"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --tb=short"
filterwarnings = ["ignore::DeprecationWarning", "ignore::UserWarning"]
